# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PESAMATRIX V2 - Cloud-to-Cloud Copy Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /strategies:
    get:
      operationId: listStrategies
      tags: [strategies]
      summary: List current user's strategies
      responses:
        "200":
          description: List of strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"
    post:
      operationId: createStrategy
      tags: [strategies]
      summary: Create a new CopyFactory strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "201":
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /strategies/{id}:
    delete:
      operationId: deleteStrategy
      tags: [strategies]
      summary: Delete a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # BINDINGS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    StrategyInput:
      type: object
      required: [strategyName, masterAccountId]
      properties:
        strategyName:
          type: string
          minLength: 2
        masterAccountId:
          type: integer
    Strategy:
      type: object
      required: [id, userId, strategyName, masterAccountId, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        copyfactoryStrategyId:
          type: ["string", "null"]
        strategyName:
          type: string
        masterAccountId:
          type: integer
        status:
          type: string
          enum: [active, paused, error]
        createdAt:
          type: string
          format: date-time

    # BINDING
...